requirements.txt
pytest==3.2.2
requests==2.18.4
vcrpy==1.11.1
